Finished adding item collection and world state changing mechanic
In this little update I added a sequence of actions where the player can pick up items, and then use them to open doors. Right now it only works for one item and one door, but I hope to be able to add duplicates that work with one another without conflict.